# document-generator-api/target/generated-sources/typescript-server/model/healthResponse.ts
/**
 * Document Generator API
 * API for generating PDF documents from templates.  This microservice provides endpoints for generating various types of insurance documents in PDF format using predefined templates and data models. 
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@autosure.co.nz
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class HealthResponse {
    /**
    * The health status of the service
    */
    'status': HealthResponse.StatusEnum;
    /**
    * ISO timestamp of the health check
    */
    'timestamp': Date;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "status",
            "baseName": "status",
            "type": "HealthResponse.StatusEnum"
        },
        {
            "name": "timestamp",
            "baseName": "timestamp",
            "type": "Date"
        }    ];

    static getAttributeTypeMap() {
        return HealthResponse.attributeTypeMap;
    }
}

export namespace HealthResponse {
    export enum StatusEnum {
        Healthy = <any> 'healthy'
    }
}